+++ backend/src/main/java/com/bil372/tour_reservation/repository/UserRepository.java
@@ -1,10 +1,13 @@
 package com.bil372.tour_reservation.repository;
 
 import com.bil372.tour_reservation.entity.User;
+
+import jakarta.persistence.criteria.CriteriaBuilder.In;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     boolean existsByEmail(String email);
